+++ crates/git/src/hosting_providers/gitee.rs
@@ -71,3 +71,128 @@ impl GitHostingProvider for Gitee {
         permalink
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_build_gitee_permalink_from_ssh_url() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@gitee.com:libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: None,
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_gitee_permalink_from_ssh_url_single_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@gitee.com:libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: Some(6..6),
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs#L7";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_gitee_permalink_from_ssh_url_multi_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@gitee.com:libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: Some(23..47),
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/editor/src/git/permalink.rs#L24-48";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_gitee_permalink_from_https_url() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://gitee.com/libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/zed/src/main.rs",
+                selection: None,
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_gitee_permalink_from_https_url_single_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://gitee.com/libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/zed/src/main.rs",
+                selection: Some(6..6),
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs#L7";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_gitee_permalink_from_https_url_multi_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "libkitten",
+            repo: "zed",
+        };
+        let permalink = Gitee.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://gitee.com/libkitten/zed.git",
+                sha: "e5fe811d7ad0fc26934edd76f891d20bdc3bb194",
+                path: "crates/zed/src/main.rs",
+                selection: Some(23..47),
+            },
+        );
+
+        let expected_url = "https://gitee.com/libkitten/zed/blob/e5fe811d7ad0fc26934edd76f891d20bdc3bb194/crates/zed/src/main.rs#L24-48";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+}
