+++ crates/git/src/hosting_providers/sourcehut.rs
@@ -73,3 +73,148 @@ impl GitHostingProvider for Sourcehut {
         permalink
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_ssh_url() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@git.sr.ht:~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: None,
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/editor/src/git/permalink.rs";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_ssh_url_with_git_prefix() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed.git",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@git.sr.ht:~rajveermalviya/zed.git",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: None,
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed.git/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/editor/src/git/permalink.rs";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_ssh_url_single_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@git.sr.ht:~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: Some(6..6),
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/editor/src/git/permalink.rs#L7";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_ssh_url_multi_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "git@git.sr.ht:~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/editor/src/git/permalink.rs",
+                selection: Some(23..47),
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/editor/src/git/permalink.rs#L24-48";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_https_url() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://git.sr.ht/~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/zed/src/main.rs",
+                selection: None,
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/zed/src/main.rs";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_https_url_single_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://git.sr.ht/~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/zed/src/main.rs",
+                selection: Some(6..6),
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/zed/src/main.rs#L7";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+
+    #[test]
+    fn test_build_sourcehut_permalink_from_https_url_multi_line_selection() {
+        let remote = ParsedGitRemote {
+            owner: "rajveermalviya",
+            repo: "zed",
+        };
+        let permalink = Sourcehut.build_permalink(
+            remote,
+            BuildPermalinkParams {
+                remote_url: "https://git.sr.ht/~rajveermalviya/zed",
+                sha: "faa6f979be417239b2e070dbbf6392b909224e0b",
+                path: "crates/zed/src/main.rs",
+                selection: Some(23..47),
+            },
+        );
+
+        let expected_url = "https://git.sr.ht/~rajveermalviya/zed/tree/faa6f979be417239b2e070dbbf6392b909224e0b/item/crates/zed/src/main.rs#L24-48";
+        assert_eq!(permalink.to_string(), expected_url.to_string())
+    }
+}
